requirements.txt
openai
rich
tiktoken
python-dotenv
cssutils